% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-split-functions.R
\name{activitysplit}
\alias{activitysplit}
\title{Split AidData Project Commitments Equally Across Activities}
\usage{
activitysplit(dat)
}
\arguments{
\item{dat}{Name of object containing AidData core data set}
}
\description{
Returns subset of proj object of rows containing sector codes in the sectors argument
}
\examples{
activitysplit(dat = dat)
}

