% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-combined-target-weights.R
\docType{data}
\name{combined_t_wts}
\alias{combined_t_wts}
\title{Matrix of weights for activity code to SDG links}
\format{}
\usage{
data(combined_t_wts)
}
\description{
Matrix of weights for activity code to SDG links
}
\examples{
data(combined_t_wts)
}
\keyword{datasets}
