% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j-wts-data.R
\docType{data}
\name{j_wts}
\alias{j_wts}
\title{Matrix of weights for activity code to SDG links}
\format{}
\usage{
data(j_wts)
}
\description{
Matrix of weights for activity code to SDG links
}
\examples{
data(j_wts)
}
\keyword{datasets}

