% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-functions.R
\name{languagecheck}
\alias{languagecheck}
\title{A Language Checking Function}
\usage{
languagecheck(text, language_list, threshold = 2/3)
}
\arguments{
\item{text}{Character string to search.}

\item{language_list}{Object containing list of words for a given language.}

\item{threshold}{Value in (0, 1] above which text is considered to be a match for the language. Set to 2/3 by default.}
}
\description{
This function checks if given percentage of text is in a certain language where the user supplies the list of words. For English, can use getEnglishWords() function.
}
\examples{
languagecheck(text = some_text, language_list = eng_words, threshold = 2/3)
}

