% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-convert-functions.R
\name{pcomms}
\alias{pcomms}
\title{A Function for Converting Documents to the AidData Flagship template (or any template, really)}
\usage{
pcomms(input_file, output_file, reference_file = "")
}
\arguments{
\item{input_file}{Name of document to convert.}

\item{output_file}{Desired name of output file.}

\item{reference_file}{Name of .dotx file to use as a reference for output style.}
}
\description{
This function is basically just pandoc but in R. Accordingly, it requires that pandoc is installed.
}
\examples{
pcomms()
}
