% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-split-functions.R
\name{projects2shape}
\alias{projects2shape}
\title{A Function for Counting World Bank IBRD-IDA Projects}
\usage{
projects2shape(shapefile, max_precision = 3, sectors = "all")
}
\arguments{
\item{shapefile}{Shapefile to append data slot with columns for financial amounts by year.}

\item{max_precision}{Maximum precision score of project locations to include.}

\item{sectors}{Sectors to include as numeric vector. Default is "all" which returns all sectors.}
}
\description{
Get count of new World Bank IBRD-IDA projects across locations by year. The returned object is the shapefile appended with a column counting new World Bank projects for every year.
}
\examples{
projects2shape(shapefile = shape, max_precision = 2, sectors = c(120, 121))
}

