% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{propover}
\alias{propover}
\title{A Function for Calculating Polygon Overlap}
\usage{
propover(p1, p2, i1, i2)
}
\arguments{
\item{p1}{First SpatialPolygons object}

\item{p2}{Second SpatialPolygons object}

\item{i1}{Index of polygon in first SpatialPolygons object}

\item{i2}{Index of polygon in second SpatialPolygons object}
}
\description{
Calculate percentage of overlap between two polygons contained in SpatialPolygons objects relative to first polygon.
}
\examples{
propover()
}

