% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-split-functions.R
\name{psplit}
\alias{psplit}
\title{A Function for Splitting AidData Projects Across Project Locations}
\usage{
psplit(proj, loc, amt_var, ID)
}
\arguments{
\item{proj}{Data frame containing projects and financial amounts}

\item{loc}{Data frame containing locations}

\item{amt_var}{Name of variable in proj containing financial amounts}

\item{ID}{Name of identifying variable in both data frames}
}
\description{
Get amount for equal splits of projects across locations
}
\examples{
psplit()
}

