% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-split-functions.R
\name{psplityear}
\alias{psplityear}
\title{A Function for Splitting AidData Projects Across Project Locations and Years}
\usage{
psplityear(proj, loc, amt_var, ID, start_var, end_var)
}
\arguments{
\item{proj}{Data frame containing projects and financial amounts}

\item{loc}{Data frame containing locations}

\item{amt_var}{Name of variable in proj containing financial amounts}

\item{ID}{Name of identifying variable in both data frames}

\item{start_var}{Name of variable in proj indicating start year of project}

\item{end_var}{Name of variable in proj indicating end year of project}
}
\description{
Get amount for equal splits of projects across locations, split across years of project duration
}
\examples{
psplityear()
}
