% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg-coding-functions.R
\name{sdg_coder}
\alias{sdg_coder}
\title{A Function for Getting SDG Estimates for Projects with AidData Activity Codes}
\usage{
sdg_coder(dat, single_activity = FALSE)
}
\arguments{
\item{dat}{Data frame containing projects and financial amounts. Name of activity codes variable assumed to be "aiddata_activity_codes" and financial amount variable "commitment_amount_usd_constant".}

\item{single_activity}{Logical value indicating whether the projects in the data frame have only a single activity. Can use activity_counter() to pre-filter data.}
}
\description{
Get amount for equal splits of projects across locations
}
\examples{
sdg_coder(dat = aiddata_core_research_release, single_activity = FALSE)
}

