% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg-coding-functions.R
\name{sdg_direct}
\alias{sdg_direct}
\title{A Function for Getting SDG Estimates using Direct Manual Coding of SDGs}
\usage{
sdg_direct(sdg_data, code_column, value_column, target = FALSE)
}
\arguments{
\item{sdg_data}{Data frame object.}

\item{code_column}{Quoted name of column containing SDG codes.}

\item{value_column}{Quoted name of column containing financial values.}

\item{target}{Logical. If TRUE, provides target-level estimates. Defaults to FALSE.}
}
\description{
This function uses the sdg_direct_single function in a loop for application to a full data set.
}
\examples{
sdg_direct()
}
