% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg-coding-functions.R
\name{sdg_direct_single}
\alias{sdg_direct_single}
\title{A Function for Getting SDG Estimates using Direct Manual Coding of SDGs}
\usage{
sdg_direct_single(codes, project_value, target_level = FALSE)
}
\arguments{
\item{codes}{Column / row entry containing SDG codes.}

\item{project_value}{Column / row entry containing SDG codes}

\item{target_level}{Logical. If TRUE, provides target-level estimates. Defaults to FALSE.}
}
\description{
Get project-level estimates of contributions to SDG goals targets using the direct coding method.
}
\examples{
sdg_direct_single(codes = data$sdg_codes[1], project_value = data$project_value[1])
}
