% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-split-functions.R
\name{sectorfilter}
\alias{sectorfilter}
\title{Filter AidData Projects by Sector Codes}
\usage{
sectorfilter(sectors, projects)
}
\arguments{
\item{sectors}{Sectors to include as numeric vector.}
}
\description{
Returns subset of proj object of rows containing sector codes in the sectors argument
}
\examples{
sectorfilter(sectors = sector, proj = data_name)
}
\keyword{Data}
\keyword{containing}
\keyword{object}
\keyword{projects}
\keyword{projects.}

