% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg-coding-functions.R
\name{target_coder}
\alias{target_coder}
\title{A Function for Getting Target-level SDG Estimates for Projects with AidData Activity Codes}
\usage{
target_coder(dat, single_activity = FALSE, coalesced_purpose = FALSE)
}
\arguments{
\item{dat}{Data frame containing projects and financial amounts. Name of activity codes variable assumed to be "aiddata_activity_codes" and financial amount variable "commitment_amount_usd_constant".}

\item{single_activity}{Logical value indicating whether all the projects in the data frame have only a single activity. Can use activity_counter() to pre-filter data. By default the function can deal with cases that have single or multiple activities, but applying the function to cases with single activities takes more time than is necessary. If TRUE, run on subset of data without multiple activity codes and bind with other data later.}

\item{coalesced_purpose}{Logical value indicating whether function should use alternative method of coalesced purpose code to SDG weights (p_wts). If TRUE, run on subset of data without activity codes and bind with other data later.}
}
\description{
Get project-level estimates of contributions to SDG targets. This function requires three input variables with specific names: aiddata_id, aiddata_activity_codes, and commitment_amount_usd_constant. If coalesced_purpose == TRUE, then there should be a column called coalesced_purpose_code.
}
\examples{
sdg_target_coder(dat = aiddata_core_research_release)
}

