% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{text_matcher}
\alias{text_matcher}
\title{A Function for Matching Text with Exclusions}
\usage{
text_matcher(x, include_text, exclude_text = NULL, ignore_case = TRUE)
}
\arguments{
\item{x}{Character vector containing text to search.}

\item{include_text}{Character vector containing terms to find.}

\item{exclude_text}{Character vector containing terms to exclude. Defaults to NULL. If exclude_text (character vector of words) is nonempty, text_matcher() first filters out any of the terms included in exclude_text before matching the include_text}

\item{ignore_case}{logical. Sets value of ignore.case in gsub and grepl.}
}
\description{
text_matcher() returns logical vector indicating whether each element in x matches terms included in include_text. If exclude_text (character vector of words) is nonempty, text_matcher() first filters out any of the terms included in exclude_text before matching the include_text
}
\examples{
textmatcher(x = c("hello", "not hello"), include_text = "hello", exclude_text = "not hello", ignore_case = TRUE)
}

