% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-functions.R
\name{tsbind}
\alias{tsbind}
\title{A Data Binding Function}
\usage{
tsbind(df, begin, end)
}
\arguments{
\item{df}{Data frame object to replicate.}

\item{begin}{Numeric starting value}

\item{end}{Numeric ending value}
}
\description{
This function turns cross-section units into cross-sectional, time-series units.
}
\examples{
tsbind(df = data_frame, begin = 2000, end = 2010)
}
