% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wts-data.R
\docType{data}
\name{wts}
\alias{wts}
\title{Matrix of weights for activity code to SDG links}
\format{}
\usage{
data(wts)
}
\description{
Matrix of weights for activity code to SDG links
}
\examples{
data(wts)
}
\keyword{datasets}

